+++ docs/examples/conf.py
@@ -1,0 +1,12 @@
+from __future__ import unicode_literals
+
+import os
+import multiprocessing
+from datetime import datetime, date
+
+home = os.getcwd()
+name = 'My App'
+dostuff = name.istitle()
+duedate = date.today()
+submodule = Subsection()
+submodule.retry = True