+++ docs/examples/defaults.py
@@ -1,0 +1,11 @@
+from datetime import date, datetime
+
+home =  '/tmp/myapp'
+name =  'MyApp'
+dostuff = False
+times = 4
+duedate = date(2014, 10, 30),
+things = ['Huey', 'Dewey', 'Louie']
+submodule = Subsection()
+submodule.retry = False
+submodule.lastrun = datetime(2014, 10, 30, 16, 40, 22)