+++ HISTORY.rst
@@ -9,7 +9,9 @@ History
 * Integration with argparse: The Commandline source now accepts an
   optional parse parameter, which should be a configured
   argparse.ArgumentParser object. Most features of argparse, such as
-  specifying the type of arguments, and automatic help text 
+  specifying the type of arguments, and automatic help text
+* A new Source, PyFile, for reading configuration from python source
+  files.
 
 0.1.0 (2014-11-03)
 ------------------